requirements.txt
Fabric==1.8.3
PyYAML==3.11